+++ src/sphinx_codeautolink/extension/warn.py
@@ -1,0 +1,5 @@
+"""Logging definitions."""
+from sphinx.util.logging import getLogger
+
+logger = getLogger('sphinx_codeautolink')
+warn_type = 'codeautolink'